move tests into their own class
